fix: failing test and lint
