+++ .prettierignore
@@ -1,3 +1,4 @@
 **/types.ts
 **/dist_electron
 **/dummy/*.json
+dbs
